fix: preserve fractional confidence in VoteMarketSentimentParams
Change confidence field from i32 to f64 to match MarketSentimentVote
response modeling and prevent lossy integer rounding on fractional values.

+++ src/types.rs
@@ -3837,7 +3837,7 @@ pub struct MarketSentimentVote {
 #[serde(rename_all = "camelCase")]
 pub struct VoteMarketSentimentParams {
     pub direction: String,
-    pub confidence: i32,
+    pub confidence: f64,
 }
 
 /// Aggregated, market-controller-derived sentiment report.
